requirements.txt
PyYAML>=6.0.1
qdrant-client>=1.9.0
psycopg[binary]>=3.1.19
rich>=13.7.1
